allow module to be used for web + native
