+++ src/formtemplates/counter.native.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { StyleSheet, Text, TextInput, TouchableOpacity, View } from 'react-native';
 
-import scstyles from '../scstyles';
+import scstyles from 'scstyles';
 
 export default function(locals) {
   var stylesheet = locals.stylesheet;
