+++ src/palette.js
@@ -1,0 +1,25 @@
+const palette = {
+  blue: '#265d79',
+  lightBlue: '#588aa8',
+  darkBlue: '#00344d',
+  orange: '#EC7B3F',
+  lightOrange: '#ffab6c',
+  darkOrange: '#b44d11',
+  green: '#209678',
+  lightGreen: '#ACCD38',
+  gray: '#F5F5F6',
+  darkGray: '#E1E2E1',
+  LABEL_COLOR: '#000000',
+  INPUT_COLOR: '#000000',
+  ERROR_COLOR: '#a94442',
+  HELP_COLOR: '#999999',
+  BORDER_COLOR: '#cccccc',
+  DISABLED_COLOR: '#777777',
+  DISABLED_BACKGROUND_COLOR: '#eeeeee',
+  FONT_SIZE: 17,
+  FONT_WEIGHT: '500',
+  INPUT_HEIGHT: 46,
+  BORDER_RADIUS: 2,
+};
+
+export default palette;
